Modified the database and modified the script.

+++ ACLager/Models/ClassAdditions/WasteReport.cs
@@ -16,8 +16,8 @@ namespace ACLager.Models {
             public System.DateTime Date { get; set; }
             [DisplayName("Mængde")]
             [Required(ErrorMessage = "Der skal angives en spildt mængde")]
-            [Range(1, long.MaxValue)]
-            public long Amount { get; set; }
+            [Range(1, Int32.MaxValue)]
+            public decimal Amount { get; set; }
             [DisplayName("Beskrivelse")]
             public string Description { get; set; }
         }
